requirements-dev.txt
black==21.7b0
chardet==4.0.0
configparser==5.2.0
coverage==6.0.1
icecream==2.1.1
Inject==4.3.1
json2xml==3.14.0
mock==4.0.3
openpyxl==3.0.7
polib==1.1.1
pydantic==1.8.2
pytest==6.2.5
python-dotenv==0.19.2
PyYAML==6.0
requests==2.26.0
tox==3.24.5
sty==1.0.0rc1
wheel>=0.38.1
XlsxWriter==3.0.1
